+++ apps/fluent-tester/src/RNTester/TestComponents/Button/ButtonFocusTest.tsx
@@ -2,10 +2,8 @@ import { Button } from '@fluentui-react-native/button';
 import { IFocusable } from '@fluentui-react-native/interactive-hooks';
 import { Stack } from '@fluentui-react-native/stack';
 import * as React from 'react';
-import { View, findNodeHandle, Text } from 'react-native';
-import { Separator } from '@fluentui-react-native/separator';
-import { stackStyle, commonTestStyles as commonStyles } from '../Common/styles';
-import { BUTTON_TESTPAGE } from './consts';
+import { findNodeHandle } from 'react-native';
+import { stackStyle } from '../Common/styles';
 
 export const ButtonFocusTest: React.FunctionComponent<{}> = () => {
   const [state, setState] = React.useState({
@@ -24,15 +22,9 @@ export const ButtonFocusTest: React.FunctionComponent<{}> = () => {
   }, [state, setState]);
 
   return (
-    <View>
-      <Text style={commonStyles.section} testID={BUTTON_TESTPAGE}>
-        Basic Buttons
-      </Text>
-      <Separator />
-      <Stack style={stackStyle}>
-        <Button content={state.focused ? 'Focused' : 'Not Focused'} componentRef={buttonRef} accessibilityLabel="overridden button name" />
-        <Button content="Click to focus" onClick={onFocus} tooltip="button tooltip" />
-      </Stack>
-    </View>
+    <Stack style={stackStyle}>
+      <Button content={state.focused ? 'Focused' : 'Not Focused'} componentRef={buttonRef} accessibilityLabel="overridden button name" />
+      <Button content="Click to focus" onClick={onFocus} tooltip="button tooltip" />
+    </Stack>
   );
 };