+++ packages/components/Button/CHANGELOG.md
@@ -1,8 +1,33 @@
 # Change Log - @fluentui-react-native/button
 
-This log was last generated on Wed, 17 Jun 2020 05:00:56 GMT and should not be manually modified.
+This log was last generated on Sat, 01 Aug 2020 00:34:34 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.10.0
+
+Sat, 01 Aug 2020 00:34:34 GMT
+
+### Minor changes
+
+- specify types for tests (krsiler@microsoft.com)
+
+## 0.9.27
+
+Tue, 28 Jul 2020 18:21:12 GMT
+
+### Patches
+
+- rename some token types and update Text (jasonmo@microsoft.com)
+
+## 0.9.26
+
+Thu, 23 Jul 2020 21:28:55 GMT
+
+### Patches
+
+- set backgroundColor on focus (lehon@microsoft.com)
+- manually update versions (taamireh@microsoft.com)
 
 ## 0.9.13