+++ packages/components/Callout/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/callout
 
-This log was last generated on Fri, 26 Jun 2020 18:58:23 GMT and should not be manually modified.
+This log was last generated on Sat, 01 Aug 2020 00:34:34 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.0
+
+Sat, 01 Aug 2020 00:34:34 GMT
+
+### Minor changes
+
+- specify types for tests (krsiler@microsoft.com)
+
+## 0.10.5
+
+Thu, 23 Jul 2020 21:28:55 GMT
+
+### Patches
+
+- manually update versions (taamireh@microsoft.com)
 
 ## 0.10.0