+++ packages/components/RadioGroup/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/radio-group
 
-This log was last generated on Wed, 17 Jun 2020 05:00:56 GMT and should not be manually modified.
+This log was last generated on Tue, 28 Jul 2020 18:21:12 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.75
+
+Tue, 28 Jul 2020 18:21:12 GMT
+
+### Patches
+
+- rename some token types and update Text (jasonmo@microsoft.com)
+
+## 0.3.74
+
+Thu, 23 Jul 2020 21:28:55 GMT
+
+### Patches
+
+- manually update versions (taamireh@microsoft.com)
 
 ## 0.3.61